.......... [ZBXNEXT-6306] changed return and internal parsing type to uint64 and added EINTR err check

+++ src/go/pkg/procfs/procfs_test.go
@@ -50,7 +50,7 @@ func Test_byteFromProcFileData(t *testing.T) {
 	tests := []struct {
 		name      string
 		args      args
-		wantValue float64
+		wantValue uint64
 		wantFound bool
 		wantErr   bool
 	}{
